More optimizations
